requirements.txt
RPi.GPIO>=0.6.5
smbus>=1.1.post2
